feat: tweaks and more caching
