+++ app/(dashboard)/home/posts/[slug]/page.tsx
@@ -1,3 +1,5 @@
+"use cache";
+
 import { getPost } from "@/utils/db";
 import PostFullView from "@/components/posts/PostFullView";
