+++ next.config.ts
@@ -3,6 +3,7 @@ import type { NextConfig } from "next";
 const nextConfig: NextConfig = {
   experimental: {
     authInterrupts: true,
+    useCache: true,
   },
   images: {
     remotePatterns: [
