+++ utils/session.ts
@@ -1,3 +1,4 @@
+import "server-only";
 import { auth } from "@/auth";
 import { redirect } from "next/navigation";
 import config from "./config";
